+++ Services/RequestServer/Request.h
@@ -29,7 +29,7 @@ namespace RequestServer {
 class Request : public HTTP::CacheRequest {
 public:
     static NonnullOwnPtr<Request> fetch(
-        i32 request_id,
+        u64 request_id,
         Optional<HTTP::DiskCache&> disk_cache,
         ConnectionFromClient& client,
         void* curl_multi,
@@ -42,7 +42,7 @@ public:
         Core::ProxyData proxy_data);
 
     static NonnullOwnPtr<Request> connect(
-        i32 request_id,
+        u64 request_id,
         ConnectionFromClient& client,
         void* curl_multi,
         Resolver& resolver,
@@ -77,7 +77,7 @@ private:
     };
 
     Request(
-        i32 request_id,
+        u64 request_id,
         Optional<HTTP::DiskCache&> disk_cache,
         ConnectionFromClient& client,
         void* curl_multi,
@@ -90,7 +90,7 @@ private:
         Core::ProxyData proxy_data);
 
     Request(
-        i32 request_id,
+        u64 request_id,
         ConnectionFromClient& client,
         void* curl_multi,
         Resolver& resolver,
@@ -120,7 +120,7 @@ private:
 
     ConnectionFromClient& client();
 
-    i32 m_request_id { 0 };
+    u64 m_request_id { 0 };
     Type m_type { Type::Fetch };
     State m_state { State::Init };
